move OHazelcastClusterMetadataManager to its own class
